+++ Revo.Infrastructure/InfrastructureConfigurationSection.cs
@@ -1,9 +1,4 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-using Revo.Core.Configuration;
+﻿using Revo.Core.Configuration;
 using Revo.Infrastructure.Events.Async;
 
 namespace Revo.Infrastructure